refactor: Unificar apps de IA
