+++ src/apps/ai_engine/services/diet_assistant.py
