+++ src/apps/ai_engine/templates/ai_engine/saved_items.html
@@ -1,0 +1,201 @@
+{% extends "base.html" %} {% block content %}
+<script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
+
+<div class="max-w-5xl mx-auto py-10 px-4 sm:px-6 lg:px-8">
+  <div class="flex justify-between items-end mb-8">
+    <div>
+      <h1 class="text-3xl font-bold text-slate-800">Meus Planos e Receitas</h1>
+      <p class="text-slate-500 mt-1">
+        Gira o conteúdo que guardou através do Assistente CalorAI.
+      </p>
+    </div>
+    <div class="flex gap-3">
+      <a
+        href="{% url 'ai-ui:shopping-list' %}"
+        class="bg-emerald-100 text-emerald-700 px-4 py-2 rounded-lg font-semibold hover:bg-emerald-200 transition"
+      >
+        🛒 Lista Automática
+      </a>
+      <a
+        href="{% url 'ai-ui:chat-page' %}"
+        class="bg-emerald-600 text-white px-4 py-2 rounded-lg font-semibold hover:bg-emerald-700 transition"
+      >
+        + Nova Conversa
+      </a>
+    </div>
+  </div>
+
+  <h2
+    class="text-xl font-bold text-emerald-700 border-b border-emerald-100 pb-2 mb-6"
+  >
+    Dietas Guardadas
+  </h2>
+
+  {% if not dietas %}
+  <p class="text-slate-400 italic mb-10">
+    Ainda não guardou nenhum plano alimentar.
+  </p>
+  {% else %}
+  <div class="grid grid-cols-1 md:grid-cols-2 gap-6 mb-12">
+    {% for dieta in dietas %}
+    <div
+      class="bg-white rounded-2xl p-6 shadow-sm border border-slate-100 flex flex-col h-full"
+    >
+      <h3 class="font-bold text-lg text-slate-800 mb-2">{{ dieta.title }}</h3>
+      <p class="text-xs text-slate-400 mb-4">
+        Guardado em: {{ dieta.created_at|date:"d M Y, H:i" }}
+      </p>
+
+      <div
+        class="prose prose-sm prose-emerald max-w-none flex-grow markdown-content overflow-y-auto max-h-60 pr-2"
+        data-content="{{ dieta.content }}"
+      ></div>
+
+      <div class="mt-4 pt-4 border-t border-slate-100 flex flex-col gap-3">
+        <form
+          action="{% url 'ai:api-edit-item-ai' %}"
+          method="POST"
+          class="flex gap-2 items-center"
+          onsubmit="mostrarCarregamento(this)"
+        >
+          {% csrf_token %}
+          <input type="hidden" name="id" value="{{ dieta.id }}" />
+          <input type="hidden" name="type" value="dieta" />
+          <input
+            type="text"
+            name="instruction"
+            placeholder="✨ Editar com IA (Ex: sem lactose)"
+            class="flex-grow px-3 py-1.5 text-sm bg-slate-50 border border-slate-200 rounded-lg focus:outline-none focus:border-emerald-500"
+            required
+          />
+          <button
+            type="submit"
+            class="text-sm font-bold text-emerald-600 hover:text-emerald-800 whitespace-nowrap transition-all"
+          >
+            Atualizar
+          </button>
+        </form>
+
+        <div class="flex justify-end">
+          <form
+            action="{% url 'ai:api-delete-item' %}"
+            method="POST"
+            onsubmit="
+              return confirm('Tem a certeza que deseja excluir esta dieta?');
+            "
+          >
+            {% csrf_token %}
+            <input type="hidden" name="id" value="{{ dieta.id }}" />
+            <input type="hidden" name="type" value="dieta" />
+            <button
+              type="submit"
+              class="text-xs font-semibold text-red-500 hover:text-red-700"
+            >
+              Excluir
+            </button>
+          </form>
+        </div>
+      </div>
+    </div>
+    {% endfor %}
+  </div>
+  {% endif %}
+
+  <h2
+    class="text-xl font-bold text-orange-600 border-b border-orange-100 pb-2 mb-6"
+  >
+    Receitas Guardadas
+  </h2>
+
+  {% if not receitas %}
+  <p class="text-slate-400 italic">Ainda não guardou nenhuma receita.</p>
+  {% else %}
+  <div class="grid grid-cols-1 md:grid-cols-2 gap-6">
+    {% for receita in receitas %}
+    <div
+      class="bg-white rounded-2xl p-6 shadow-sm border border-slate-100 flex flex-col h-full"
+    >
+      <h3 class="font-bold text-lg text-slate-800 mb-2">{{ receita.title }}</h3>
+      <p class="text-xs text-slate-400 mb-4">
+        Guardado em: {{ receita.created_at|date:"d M Y, H:i" }}
+      </p>
+
+      <div
+        class="prose prose-sm prose-orange max-w-none flex-grow markdown-content overflow-y-auto max-h-60 pr-2"
+        data-content="{{ receita.content }}"
+      ></div>
+
+      <div class="mt-4 pt-4 border-t border-slate-100 flex flex-col gap-3">
+        <form
+          action="{% url 'ai:api-edit-item-ai' %}"
+          method="POST"
+          class="flex gap-2 items-center"
+          onsubmit="mostrarCarregamento(this)"
+        >
+          {% csrf_token %}
+          <input type="hidden" name="id" value="{{ receita.id }}" />
+          <input type="hidden" name="type" value="receita" />
+          <input
+            type="text"
+            name="instruction"
+            placeholder="✨ Editar com IA (Ex: trocar frango por atum)"
+            class="flex-grow px-3 py-1.5 text-sm bg-slate-50 border border-slate-200 rounded-lg focus:outline-none focus:border-orange-500"
+            required
+          />
+          <button
+            type="submit"
+            class="text-sm font-bold text-orange-500 hover:text-orange-700 whitespace-nowrap transition-all"
+          >
+            Atualizar
+          </button>
+        </form>
+
+        <div class="flex justify-end">
+          <form
+            action="{% url 'ai:api-delete-item' %}"
+            method="POST"
+            onsubmit="
+              return confirm('Tem a certeza que deseja excluir esta receita?');
+            "
+          >
+            {% csrf_token %}
+            <input type="hidden" name="id" value="{{ receita.id }}" />
+            <input type="hidden" name="type" value="receita" />
+            <button
+              type="submit"
+              class="text-xs font-semibold text-red-500 hover:text-red-700"
+            >
+              Excluir
+            </button>
+          </form>
+        </div>
+      </div>
+    </div>
+    {% endfor %}
+  </div>
+  {% endif %}
+</div>
+
+<script>
+  function mostrarCarregamento(formulario) {
+    const botao = formulario.querySelector('button[type="submit"]');
+    const inputTexto = formulario.querySelector('input[type="text"]');
+
+    botao.innerHTML = "⏳ A processar...";
+    botao.classList.add("opacity-50", "pointer-events-none");
+
+    inputTexto.classList.add("opacity-50", "bg-slate-100");
+    inputTexto.readOnly = true;
+  }
+
+  document.addEventListener("DOMContentLoaded", function () {
+    const markdownDivs = document.querySelectorAll(".markdown-content");
+    markdownDivs.forEach((div) => {
+      const rawContent = div.getAttribute("data-content");
+      if (rawContent) {
+        div.innerHTML = marked.parse(rawContent);
+      }
+    });
+  });
+</script>
+{% endblock %}
