+++ src/apps/ai_engine/ui_urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+
+from . import views
+
+app_name = "ai-ui"
+
+urlpatterns = [
+    path("chat/", views.chat_page, name="chat-page"),
+    path("salvos/", views.saved_items_page, name="saved-items"),
+    path("lista-compras/", views.shopping_list_page, name="shopping-list"),
+]
