--- src/apps/assistant/templates/assistant/chat.html
@@ -1,152 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
-
-<div class="min-h-[85vh] py-8 px-4 sm:px-6 lg:px-8 flex flex-col">
-  <div class="max-w-4xl mx-auto w-full flex-grow flex flex-col bg-white rounded-[2rem] shadow-sm border border-slate-100 overflow-hidden">
-
-    <div class="bg-emerald-600 px-6 py-4 flex items-center gap-3 border-b border-emerald-700">
-      <div class="w-10 h-10 rounded-full bg-white text-emerald-600 flex items-center justify-center font-bold text-xl">AI</div>
-      <div>
-        <h1 class="text-lg font-bold text-white">Assistente CalorAI</h1>
-        <p class="text-emerald-100 text-xs">A analisar o seu perfil...</p>
-      </div>
-
-      <div class="ml-auto">
-          <a href="{% url 'assistant:saved-items' %}" class="bg-emerald-700 hover:bg-emerald-800 text-white text-sm font-bold py-2 px-4 rounded-lg transition shadow-sm border border-emerald-600">
-              📂 Meus Planos e Receitas
-          </a>
-      </div>
-    </div>
-    <div id="chat-history" class="flex-grow p-6 overflow-y-auto space-y-6 bg-slate-50 flex flex-col" style="max-height: 60vh;">
-      <div class="flex justify-start">
-        <div class="bg-white border border-slate-200 text-slate-800 px-5 py-4 rounded-2xl rounded-tl-sm max-w-[85%] shadow-sm text-sm">
-          Olá! Sou o seu assistente nutricional inteligente. Baseado na sua meta e restrições, o que gostaria de planejar para a próxima refeição?
-        </div>
-      </div>
-    </div>
-
-    <div class="p-4 bg-white border-t border-slate-100">
-      <form id="ai-chat-form" class="flex gap-3">
-        {% csrf_token %}
-        <input type="text" id="user-message" placeholder="Ex: Sugere-me um almoço rico em proteína..." class="flex-grow px-5 py-3 bg-slate-50 border border-slate-200 focus:ring-emerald-500 rounded-xl outline-none" required autocomplete="off"/>
-        <button type="submit" id="send-btn" class="px-6 py-3 font-bold text-white bg-emerald-600 hover:bg-emerald-700 rounded-xl disabled:opacity-50">Enviar</button>
-      </form>
-    </div>
-  </div> </div>
-
-<script>
-  // Função para salvar a dieta ou receita
-  async function salvarConteudo(tipo, textoMarkdown, btnElement) {
-    const csrf = document.querySelector('[name=csrfmiddlewaretoken]').value;
-
-    // Altera o texto do botão para mostrar que está a pensar
-    const textoOriginal = btnElement.innerHTML;
-    btnElement.innerHTML = '⏳ A guardar...';
-    btnElement.disabled = true;
-
-    try {
-        const response = await fetch("{% url 'assistant:api-save-content' %}", {
-            method: 'POST',
-            headers: {'Content-Type': 'application/json', 'X-CSRFToken': csrf},
-            body: JSON.stringify({ type: tipo, content: textoMarkdown })
-        });
-
-        const data = await response.json();
-
-        if (response.ok) {
-            // Sucesso! Transforma o botão num selo visual e remove a ação de clique
-            btnElement.innerHTML = '✅ Guardado no seu perfil';
-            btnElement.classList.replace('hover:bg-emerald-200', 'cursor-default');
-            btnElement.classList.replace('hover:bg-orange-200', 'cursor-default');
-            btnElement.classList.remove('bg-emerald-100', 'text-emerald-700', 'bg-orange-100', 'text-orange-700');
-            btnElement.classList.add('bg-slate-100', 'text-slate-500');
-            btnElement.onclick = null; // Remove a função de clique
-        } else {
-            alert('Erro ao salvar: ' + (data.error || 'Erro desconhecido.'));
-            btnElement.innerHTML = textoOriginal; // Volta ao normal se der erro
-            btnElement.disabled = false;
-        }
-    } catch (error) {
-        console.error('Erro:', error);
-        alert('Falha na comunicação com o servidor.');
-        btnElement.innerHTML = textoOriginal;
-        btnElement.disabled = false;
-    }
-  }
-
-  document.getElementById('ai-chat-form').addEventListener('submit', async function(e) {
-    e.preventDefault();
-    const input = document.getElementById('user-message');
-    const btn = document.getElementById('send-btn');
-    const history = document.getElementById('chat-history');
-    const text = input.value.trim();
-    const csrf = document.querySelector('[name=csrfmiddlewaretoken]').value;
-
-    if (!text) return;
-
-    history.insertAdjacentHTML('beforeend', `<div class="flex justify-end"><div class="bg-emerald-600 text-white px-5 py-4 rounded-2xl rounded-tr-sm max-w-[85%] text-sm">${text}</div></div>`);
-    input.value = ''; btn.disabled = true; history.scrollTop = history.scrollHeight;
-
-    const loadId = 'loading-' + Date.now();
-    history.insertAdjacentHTML('beforeend', `<div id="${loadId}" class="flex justify-start"><div class="bg-white border border-slate-200 text-slate-500 px-5 py-4 rounded-2xl rounded-tl-sm text-sm">A processar...</div></div>`);
-    history.scrollTop = history.scrollHeight;
-
-    try {
-        const response = await fetch("{% url 'assistant:api-chat' %}", {
-            method: 'POST',
-            headers: {'Content-Type': 'application/json', 'X-CSRFToken': csrf},
-            body: JSON.stringify({ message: text })
-        });
-
-        const data = await response.json();
-        const loadingEl = document.getElementById(loadId);
-        if (loadingEl) loadingEl.remove();
-
-        if (response.ok) {
-            let saveBtnHTML = '';
-
-            // CORREÇÃO: Codifica o texto para ser 100% seguro contra aspas e quebras de linha no HTML
-            const safeReply = encodeURIComponent(data.reply).replace(/'/g, "%27");
-
-            // Renderiza botão apenas se não for chat normal
-            if (data.type === 'dieta') {
-                saveBtnHTML = `
-                <div class="mt-4 pt-3 border-t border-slate-100 flex justify-end">
-                    <button onclick="salvarConteudo('dieta', decodeURIComponent('${safeReply}'), this)" class="text-sm px-4 py-2 bg-emerald-100 text-emerald-700 rounded-lg font-bold hover:bg-emerald-200 transition">
-                        💾 Salvar Dieta
-                    </button>
-                </div>`;
-            } else if (data.type === 'receita') {
-                saveBtnHTML = `
-                <div class="mt-4 pt-3 border-t border-slate-100 flex justify-end">
-                    <button onclick="salvarConteudo('receita', decodeURIComponent('${safeReply}'), this)" class="text-sm px-4 py-2 bg-orange-100 text-orange-700 rounded-lg font-bold hover:bg-orange-200 transition">
-                        🍳 Salvar Receita
-                    </button>
-                </div>`;
-            }
-
-            history.insertAdjacentHTML('beforeend', `
-                <div class="flex justify-start">
-                    <div class="bg-white border border-slate-200 text-slate-800 px-6 py-5 rounded-2xl rounded-tl-sm max-w-[95%] sm:max-w-[85%] shadow-sm">
-                        <div class="prose prose-emerald prose-sm max-w-none">
-                            ${marked.parse(data.reply)}
-                        </div>
-                        ${saveBtnHTML}
-                    </div>
-                </div>
-            `);
-        } else {
-            throw new Error(data.detail || "Erro de comunicação com a IA.");
-        }
-    } catch (error) {
-        const loadingEl = document.getElementById(loadId);
-        if (loadingEl) loadingEl.remove();
-
-        history.insertAdjacentHTML('beforeend', `<div class="flex justify-start"><div class="bg-red-50 border border-red-200 text-red-700 px-5 py-4 rounded-2xl rounded-tl-sm shadow-sm text-sm">Desculpe, ocorreu um erro: ${error.message}</div></div>`);
-    } finally {
-        btn.disabled = false; input.focus(); history.scrollTop = history.scrollHeight;
-    }
-  });
-</script>
-{% endblock %}
